RAG/ragAgent/pyproject.toml
[project]
name = "ragagent"
version = "0.1.0"
description = "RAG Agent with MCP tool capabilities for document search, database queries, and semantic retrieval"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "httpx>=0.28.1",
    "openai>=1.40.0",
    "langchain>=0.2.11",
    "langchain-community>=0.2.10",
    "langchain-core>=0.2.11",
    "langchain-openai>=0.1.0",
    "langchain-chroma>=0.1.0",
    "langgraph>=0.3.18",
    "chromadb>=0.5.5",
    "pypdf>=4.2.0",
    "tiktoken>=0.7.0",
    "reportlab>=4.2.2",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "langchain_mcp_adapters",
    "mcp",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
